gemini-projects/requirements.txt
pandas
numpy
yfinance
ccxt
requests
python-dotenv
google-generativeai
pydantic
